requirements.txt
appdirs==1.4.4
attrs==21.2.0
black==21.5b1
certifi==2021.5.30
chardet==4.0.0
click==8.0.1
colorama==0.4.4
docker==5.0.0
gitdb==4.0.7
GitPython==3.1.18
hjson==3.0.2
idna==2.10
iniconfig==1.1.1
isort==5.8.0
mypy-extensions==0.4.3
packaging==20.9
pathspec==0.8.1
pluggy==0.13.1
py==1.10.0
pyparsing==2.4.7
pytest==6.2.4
pytest-sugar==0.9.4
regex==2021.4.4
requests==2.25.1
six==1.16.0
smmap==4.0.0
termcolor==1.1.0
toml==0.10.2
urllib3==1.26.6
websocket-client==1.1.0
